# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级后台管理系统接口文档，覆盖认证、用户、角色、菜单、部门、字典和审计等业务模块。",
        "title": "Go Server 后台管理 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/system/dicts/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按字典类型、标签、状态等条件筛选字典数据项。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "查询字典数据列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型 ID",
                        "name": "type_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典类型编码",
                        "name": "type_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典标签",
                        "name": "label",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态：0 禁用，1 启用",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_dict.DictDataResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "在指定字典类型下新增一个字典数据项。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "创建字典数据",
                "parameters": [
                    {
                        "description": "字典数据创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_dict.CreateDictDataRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dict.DictDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/dicts/items/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据字典数据 ID 查询详细信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "获取字典数据详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dict.DictDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改字典数据项的标签、值、排序、状态和备注。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "更新字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "字典数据更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_dict.UpdateDictDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dict.DictDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据字典数据 ID 删除字典数据项。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "删除字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/dicts/lookup/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据字典类型编码返回对应的可用字典数据项。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "按类型编码查询字典数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典类型编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_dict.DictDataResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/dicts/types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按名称或编码筛选字典类型。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "查询字典类型列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典类型名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典类型编码",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_dict.DictTypeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的字典类型，用于归类字典数据项。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典管理"
                ],
                "summary": "创建字典类型",
                "parameters": [
                    {
                        "description": "字典类型创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_dict.CreateDictTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dict.DictTypeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-server_internal_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_modules_dict.CreateDictDataRequest": {
            "type": "object",
            "required": [
                "label",
                "type_id",
                "value"
            ],
            "properties": {
                "label": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "启用"
                },
                "remark": {
                    "type": "string",
                    "example": "启用状态"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "type_id": {
                    "type": "integer",
                    "example": 1
                },
                "value": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "1"
                }
            }
        },
        "internal_modules_dict.CreateDictTypeRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "user_status"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "用户状态"
                },
                "remark": {
                    "type": "string",
                    "example": "用户启用禁用状态"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                }
            }
        },
        "internal_modules_dict.DictDataResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-04-11T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "label": {
                    "type": "string",
                    "example": "启用"
                },
                "remark": {
                    "type": "string",
                    "example": "启用状态"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "type_code": {
                    "type": "string",
                    "example": "user_status"
                },
                "type_id": {
                    "type": "integer",
                    "example": 1
                },
                "value": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "internal_modules_dict.DictTypeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "user_status"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-11T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "用户状态"
                },
                "remark": {
                    "type": "string",
                    "example": "用户启用禁用状态"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_modules_dict.UpdateDictDataRequest": {
            "type": "object",
            "required": [
                "label",
                "type_id",
                "value"
            ],
            "properties": {
                "label": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "启用"
                },
                "remark": {
                    "type": "string",
                    "example": "启用状态"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "type_id": {
                    "type": "integer",
                    "example": 1
                },
                "value": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "1"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请填写 Bearer Token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}